Fix for general exceptions.
When a bulk update fails due to some other error it would cause the
bulk process to go down, the queue would then fill up because no
process was draining it, and the result would be a dead-lock because
nothing new would go in the queue.

Fix this by catching generic exceptions and printing them.