Add content to help tab.
This is done by opening the README.md for pyDat (not the one at the root of the
tree) and passing it through the markdown python module as a template tag.

It requires that pydat be listed in INSTALLED_APPS in settings.py.

While here, cleanup the pydat/README.md contents to be better formatted and not
use HTML tags.

Also, document the requirement for the markdown module in the top-of-tree
README.